extract 'schemas' in a different file
