+++ app/routes.py
@@ -1,9 +1,8 @@
 from flask import request, Response
 from sqlalchemy.orm.exc import NoResultFound
-from marshmallow import Schema, fields, ValidationError, post_load
+from marshmallow import ValidationError
 from src import utils
-from app import db, app
-from app import ma
+from app import db, app, ma
 from app.exceptions import InvalidRequestException
 from app.models import (
     Resident,
@@ -14,91 +13,15 @@ from app.models import (
     Contributor,
     HealthMutual
 )
-
-
-#   ____   ____ _   _ _____ __  __    _    ____
-#  / ___| / ___| | | | ____|  \/  |  / \  / ___|
-#  \___ \| |   | |_| |  _| | |\/| | / _ \ \___ \
-#   ___) | |___|  _  | |___| |  | |/ ___ \ ___) |
-#  |____/ \____|_| |_|_____|_|  |_/_/   \_\____/
-
-
-
-class SchemaMixin(object):
-    @post_load
-    def make_object(self, data, **kwargs):
-
-        return self.Meta.model(**data)
-
-
-def must_not_be_blank(data):
-    if not data:
-        raise ValidationError("Data not provided.")
-
-
-class PersonSchema(SchemaMixin, ma.SQLAlchemyAutoSchema):
-    class Meta:
-        include_fk = True
-        model = Person
-
-
-class CitySchema(SchemaMixin, ma.SQLAlchemyAutoSchema):
-    class Meta:
-        include_fk = True
-        model = City
-
-
-class ContributorSchema(SchemaMixin, ma.SQLAlchemyAutoSchema):
-    class Meta:
-        include_fk = True
-        model = Contributor
-
-
-class HealthMutualSchema(SchemaMixin, ma.SQLAlchemyAutoSchema):
-    class Meta:
-        include_fk = True
-        model = HealthMutual
-
-
-class ResidentSchema(SchemaMixin, ma.SQLAlchemyAutoSchema):
-    class Meta:
-        include_fk = True
-        model = Resident
-    person = fields.Nested(PersonSchema)
-    city = fields.Nested(CitySchema)
-    healthMutual = fields.Nested(HealthMutualSchema)
-    doctor = fields.Nested(PersonSchema)
-    psychiatrist = fields.Nested(PersonSchema)
-
-
-class EmergencyRelationshipSchema(SchemaMixin, ma.SQLAlchemyAutoSchema):
-    class Meta:
-        include_fk = True
-        model = EmergencyRelationship
-
-
-class ContributionRelationshipSchema(SchemaMixin, ma.SQLAlchemyAutoSchema):
-    class Meta:
-        include_fk = True
-        model = ContributionRelationship
-
-
-person_schema = PersonSchema()
-persons_schema = PersonSchema(many=True)
-resident_schema = ResidentSchema()
-residents_schema = ResidentSchema(many=True)
-city_schema = CitySchema()
-cities_schema = CitySchema(many=True)
-contributor_schema = ContributorSchema()
-contributors_schema = ContributorSchema(many=True)
-health_mutual_schema = HealthMutualSchema()
-health_mutuals_schema = HealthMutualSchema(many=True)
-emergency_relationship_schema = EmergencyRelationshipSchema()
-emergencyRelationships_schema = EmergencyRelationshipSchema(many=True)
-contribution_relationship_schema = ContributionRelationshipSchema()
-contributionRelationships_schema = ContributionRelationshipSchema(many=True)
-
-
+from app.schemas import (
+    resident_schema, residents_schema,
+    person_schema, persons_schema,
+    city_schema, cities_schema,
+    contributor_schema, contributors_schema,
+    emergency_relationship_schema, emergencyRelationships_schema,
+    contribution_relationship_schema, contribution_relationships_schema,
+    health_mutual_schema, health_mutuals_schema
+)
 #      _    ____ ___
 #     / \  |  _ \_ _|
 #    / _ \ | |_) | |
@@ -232,8 +155,8 @@ def create_new_item(
     result = schema.dump(model.query.get(item.id))
     return utils.http_response(utils.HTTPStatus.CREATED, result)
 
-# API routes
 
+# API routes
 
 @app.route("/persons", methods=["GET", "POST"])
 def person_collection() -> utils.Response:
@@ -362,7 +285,7 @@ def contributionRelationships_collection(id: str) -> utils.Response:
     if request.method == "GET":
         er_collection = ContributionRelationship.query.filter_by(
             residentId=id).all()
-        list_result = contributionRelationships_schema.dump(er_collection)
+        list_result = contribution_relationships_schema.dump(er_collection)
         return utils.http_response(utils.HTTPStatus.OK, list_result)
 
     if request.method == "POST":
@@ -392,5 +315,3 @@ def reset_db() -> utils.Response:
     db.create_all()
 
     return utils.http_response(utils.HTTPStatus.NO_CONTENT, None)
-
-
